qldb_streaming_sample/requirements.txt
amazon.ion==0.5.0
aws-kinesis-agg==1.1.2